Kill msys2_shell.cmd straggler processes
After msys2_shell.cmd runs its shell session remains which
causes the buildkite-agent to hang at the end of the job.
This kills the shell session's processes which allows the shell
to exit. This mimics what other projects (such as chocolatey)
do when running msys2_shell.cmd.

It is safe to have multiple concurrent jobs run this kill process
because we only kill processes that have a "Path" that matches
the current build's path.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>